update tf output dtype to hold fp32 for mixed precision

+++ tfdet/core/loss/segmentation.py
@@ -6,7 +6,7 @@ def dice(y_true, y_pred, smooth = 1., weight = None, reduce = True):
     n_true_class = tf.shape(y_true)[-1]
     n_pred_class = tf.shape(y_pred)[-1]
     
-    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], false_fn = lambda: y_true), y_pred.dtype)
+    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.cast(tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], y_true.dtype), false_fn = lambda: y_true), y_pred.dtype)
     y_pred = y_pred / (tf.reduce_sum(y_pred, axis = -1, keepdims = True) + tf.keras.backend.epsilon())
     y_pred = tf.clip_by_value(y_pred, tf.keras.backend.epsilon(), 1. - tf.keras.backend.epsilon())
     
@@ -38,7 +38,7 @@ def tversky(y_true, y_pred, smooth = 1., alpha = 0.5, beta = 0.5, weight = None,
     n_true_class = tf.shape(y_true)[-1]
     n_pred_class = tf.shape(y_pred)[-1]
     
-    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], false_fn = lambda: y_true), y_pred.dtype)
+    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.cast(tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], y_true.dtype), false_fn = lambda: y_true), y_pred.dtype)
     y_pred = y_pred / (tf.reduce_sum(y_pred, axis = -1, keepdims = True) + tf.keras.backend.epsilon())
     y_pred = tf.clip_by_value(y_pred, tf.keras.backend.epsilon(), 1. - tf.keras.backend.epsilon())
     
@@ -70,7 +70,7 @@ def iou_pixcel(y_true, y_pred, smooth = 1., weight = None, reduce = True):
     n_true_class = tf.shape(y_true)[-1]
     n_pred_class = tf.shape(y_pred)[-1]
     
-    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], false_fn = lambda: y_true), y_pred.dtype)
+    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.cast(tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], y_true.dtype), false_fn = lambda: y_true), y_pred.dtype)
     y_pred = y_pred / (tf.reduce_sum(y_pred, axis = -1, keepdims = True) + tf.keras.backend.epsilon())
     y_pred = tf.clip_by_value(y_pred, tf.keras.backend.epsilon(), 1. - tf.keras.backend.epsilon())
     
@@ -92,7 +92,7 @@ def generalized_dice(y_true, y_pred, smooth = 1, weight = None, reduce = True):
     n_true_class = tf.shape(y_true)[-1]
     n_pred_class = tf.shape(y_pred)[-1]
     
-    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], false_fn = lambda: y_true), y_pred.dtype)
+    y_true = tf.cast(tf.cond(tf.logical_and(tf.equal(n_true_class, 1), tf.not_equal(n_pred_class, 1)), true_fn = lambda: tf.cast(tf.one_hot(tf.cast(y_true, tf.int32), n_pred_class)[:, 0], y_true.dtype), false_fn = lambda: y_true), y_pred.dtype)
     y_pred = y_pred / (tf.reduce_sum(y_pred, axis = -1, keepdims = True) + tf.keras.backend.epsilon())
     y_pred = tf.clip_by_value(y_pred, tf.keras.backend.epsilon(), 1. - tf.keras.backend.epsilon())
